# Improve Array#shift performance on v8 > 7.1 (Node 12+)