… benchmark-ips/bm_array_shift.rb