### Command

```
$ bundle exec rake bench:ips FILE=benchmark-ips/bm_array_shift.rb
```

### Before

#### Node 10

```
=== Files: benchmark-ips/bm_array_shift.rb
bundle exec opal --dynamic-require ignore --missing-require ignore -rnodejs -ropal/platform -gbenchmark-ips -rbenchmark/ips -A benchmark-ips/bm_array_shift.rb
""
Warming up --------------------------------------
        shift no arg    39.327k i/100ms
Calculating -------------------------------------
        shift no arg    979.605k (± 1.6%) i/s -      4.916M in   5.019468s
```

#### Node 12

```

=== Files: benchmark-ips/bm_array_shift.rb
bundle exec opal --dynamic-require ignore --missing-require ignore -rnodejs -ropal/platform -gbenchmark-ips -rbenchmark/ips -A benchmark-ips/bm_array_shift.rb
""
Warming up --------------------------------------
        shift no arg     2.364k i/100ms
Calculating -------------------------------------
        shift no arg     24.879k (± 0.5%) i/s -    125.292k in   5.036277s
```

### After

#### Node 10

```
=== Files: benchmark-ips/bm_array_shift.rb
bundle exec opal --dynamic-require ignore --missing-require ignore -ropal/platform -gbenchmark-ips -rbenchmark/ips -A benchmark-ips/bm_array_shift.rb
Warming up --------------------------------------
        shift no arg    26.398k i/100ms
Calculating -------------------------------------
        shift no arg    449.226k (± 4.4%) i/s -      2.244M in   5.005441s
```

#### Node 12

```
=== Files: benchmark-ips/bm_array_shift.rb
bundle exec opal --dynamic-require ignore --missing-require ignore -ropal/platform -gbenchmark-ips -rbenchmark/ips -A benchmark-ips/bm_array_shift.rb
Warming up --------------------------------------
        shift no arg    37.305k i/100ms
Calculating -------------------------------------
        shift no arg    611.658k (± 1.0%) i/s -      3.059M in   5.001669s
```

### Summary

With this change, the code is now 2.18x times slower when using Node 10 but 24.58x times faster when using Node 12.

We cannot beat the highly optimized code produced by v8 pre 7.1 (Node 10) but the code is now "only" 1.6x times slower when using a recent version of v8/Node.

Please note that the performance on Node 14 is roughly equivalent to Node 12.

### Further improvement

The `Array#shift()` (JavaScript) function is also used in [`String#tr`](https://github.com/opal/opal/blob/53593f7a82377deedd2d8082a92efd9b6848f1e0/opal/corelib/string.rb#L1388) and [`String#tr_s`](https://github.com/opal/opal/blob/53593f7a82377deedd2d8082a92efd9b6848f1e0/opal/corelib/string.rb#L1533). It might be worth to replace this call by the optimized `shiftNoArg` function? In this case, where should I declare the `shiftNoArg` function? in `runtime.js`?

I also found an optimization for `splice` when the second argument is `1` (ie. `list.splice(position, 1)`). I found 8 occurrences in the code where the second argument is explicitly 1. When a variable is used as the second argument it might be worth to check if the value is actually 1 to use the optimized code (I think it's only worth if the array is large enough but that means that we need to add an additional check that will slow down the execution).