@Mogztter 2020-10-11
@elia I found a more generic implementation:

```js
function spliceGen (list, start, deleteCount) {
  if (deleteCount <= 0) {
    return
  }
  const length = list.length
  for (var i = start, k = i + deleteCount, n = length; k < n; i += 1, k += 1)
    list[i] = list[k]

  list.length = i
  // or:
  //for (;i < length; i++)
  //  list.pop()
}
```

Depending on the list length, calling `list.pop()` can be faster than updating the list length. The goal is to remove "trailing" elements.