@Mogztter 2020-10-12
@elia I will keep the focus of this pull request to just `Array#shift` and open a new pull request to generalize the usage of `spliceOne` / `spliceGen`.