@Mogztter 2020-10-22
@elia I've added a changelog entry and rebased/squashed my work. Let me know if there's anything left to do.